refactor: use common files
