+++ src/Kaonavi.NET.Core/Entities/WebhookConfig.cs
@@ -23,17 +23,3 @@ public record WebhookConfigPayload(Uri Url, IReadOnlyList<WebhookEvent> Events, 
 /// <param name="SecretToken"><inheritdoc cref="WebhookConfigPayload" path="/param[@name='SecretToken']"/></param>
 public record WebhookConfig(int Id, Uri Url, IReadOnlyList<WebhookEvent> Events, string SecretToken)
     : WebhookConfigPayload(Url, Events, SecretToken);
-
-/// <summary>
-/// Webhookで利用できるイベント
-/// </summary>
-[JsonConverter(typeof(JsonStringEnumConverter<WebhookEvent>))]
-public enum WebhookEvent
-{
-    /// <summary>メンバーが新しく登録された</summary>
-    [JsonStringEnumMemberName("member_created")] MemberCreated = 1,
-    /// <summary>メンバーの基本情報が更新された</summary>
-    [JsonStringEnumMemberName("member_updated")] MemberUpdated = 2,
-    /// <summary>メンバーが削除された</summary>
-    [JsonStringEnumMemberName("member_deleted")] MemberDeleted = 3,
-}
