+++ src/Kaonavi.NET.Server/Kaonavi.NET.Server.csproj
@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFrameworks>net8.0;net6.0</TargetFrameworks>
+    <TargetFrameworks>net9.0;net8.0</TargetFrameworks>
     <RootNamespace>Kaonavi.Net.Server</RootNamespace>
 
     <Description>Server side parser for Kaonavi Webhook</Description>
@@ -16,10 +16,12 @@
 
   <ItemGroup>
     <None Include="./README.md" Pack="true" PackagePath="/" />
-    <Compile Include="..\Kaonavi.Net.Core\Json\DateTimeConverter.cs" />
+    <Compile Include="../Kaonavi.Net.Core/Entities/WebhookEvent.cs" />
+    <Compile Include="../Kaonavi.Net.Core/Json/DateTimeConverter.cs" />
   </ItemGroup>
 
   <ItemGroup>
     <FrameworkReference Include="Microsoft.AspNetCore.App" />
+    <PackageReference Include="System.Text.Json" Version="9.0.0" Condition="'$(TargetFramework)'!='net9.0'" />
   </ItemGroup>
 </Project>
